fix(user-spending-model): change route for GetTransactionsByUserSpendingModelAsync
