+++ MoneyEz.API/Controllers/TransactionsController.cs
@@ -29,12 +29,6 @@ namespace MoneyEz.API.Controllers
         {
             return ValidateAndExecute(() => _transactionService.GetAllTransactionsForUserAsync(paginationParameter));
         }
-        
-        [HttpGet("user-spending-model/{userSpendingModelId}")]
-        public Task<IActionResult> GetAllTransactionsBySpendingModel([FromQuery] PaginationParameter paginationParameter, Guid userSpendingModelId)
-        {
-            return ValidateAndExecute(() => _transactionService.GetTransactionsByUserSpendingModelAsync(paginationParameter, userSpendingModelId));
-        }
 
         [Authorize(Roles = nameof(RolesEnum.USER))]
         [HttpGet("user/{transactionId}")]
