+++ MoneyEz.Services/Services/Interfaces/ITransactionService.cs
@@ -17,7 +17,5 @@ namespace MoneyEz.Services.Services.Interfaces
         Task<BaseResultModel> DeleteTransactionAsync(Guid transactionId);
         Task<BaseResultModel> GetAllTransactionsForAdminAsync(PaginationParameter paginationParameter);
         Task<BaseResultModel> GetTransactionByGroupIdAsync(PaginationParameter paginationParameter, TransactionFilter transactionFilter);
-        Task<BaseResultModel> GetTransactionsByUserSpendingModelAsync(PaginationParameter paginationParameter, Guid userSpendingModelId);
-
     }
 }
